feat: expanse table on report page

+++ database/migrations/2025_12_24_111717_create_expenses_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
       $table->id();
       $table->enum('type', ['bayar it', 'belanja', 'lain-lain']);
       $table->string('description');
-      $table->integer('amount');
+      $table->decimal('amount', 15, 2);
       $table->timestamps();
     });
   }
